requirements.txt
Django==4.0.4
coverage==5.3.1
djangorestframework==3.12.2
uritemplate==3.0.1
psycopg2-binary==2.9.3
celery==5.2.6
django-environ==0.8.1
beautifulsoup4==4.11.1
requests==2.22.0
django-cors-headers==3.13.0
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.7.1
django-material-admin==1.8.6
django-phonenumbers==1.0.1
django-phonenumber-field==7.0.0
pandas==2.0.2
openpyxl==3.1.2
